Cifar_10/requirements.txt
streamlit==1.25.0
numpy==1.24.3
tensorflow==2.13.0
Pillow==9.4.0
